+++ test/New-C4ContainerBoundary.Tests.ps1
@@ -1,13 +1,13 @@
 #Requires -Modules @{ ModuleName='Pester'; ModuleVersion='5.0.0' }
 
-Describe New-ContainerBoundary {
+Describe New-C4ContainerBoundary {
 
     BeforeAll {
         Import-Module $PSScriptRoot/../src/PsMermaidTools.psd1 -Force
     }
 
     It works {
-        $boundary = New-MermaidContainerBoundary -Key A -Name foo
+        $boundary = New-MermaidC4ContainerBoundary -Key A -Name foo
         $boundary | Should -Not -BeNullOrEmpty
         $boundary.Key | Should -Be A
         $boundary.Name | Should -Be foo