+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - new diagram type flowcharts.
-- new command `Add-Component`.
+- new command `Add-C4Component`.
 - new command `Add-C4ContainerBoundary`.
 - new command `New-C4Component`.
+- new command `New-C4ContainerBoundary`.
 
 ### Changed